+++ src/pages/authorization/login.tsx
@@ -5,6 +5,7 @@ import {authUser, authWithError} from './store';
 import {IFormInputs} from '@shared/forms/interfaces/interfaces';
 import LoginForm from '@entities/loginForm/loginForm';
 import useLocalStorage from '@shared/hooks/useLocalStorage';
+import {loginWithLoginAndPassword} from '@shared/api/services/authorizationService';
 
 function Login() {
   const onSubmit= (data: IFormInputs) => loginHandler(data.login, data.password);
