+++ src/shared/api/services/todosService/fetchTodos.ts
@@ -1,12 +1,6 @@
 import {ITodo} from '@shared/interfaces';
+import {fetchRequest} from '@shared/api/services/constants';
 
-const BASE_URL = 'http://localhost:4444/api/'
-
-export const fetchRequest = (url: string, options: RequestInit = {}) => new Promise((resolve, reject) => {
-  fetch(`${BASE_URL}${url}`, options)
-      .then((res) => resolve(res))
-      .catch((error) => reject(error))
-})
 
 export const getUserTodos = (userId: string) => new Promise((resolve, reject) => {
   fetchRequest(`get_all/${userId}`)
